requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt<4
python-multipart
pydantic
pydantic[email]
pydantic-settings
python-dotenv
httpx
scikit-learn
numpy
pandas
openai
requests